Move _convert_base_to_offset to pdcompat and add a few more tests
